add page forget and resetand verification code

+++ src/components/auth/AuthRegister.jsx
@@ -7,16 +7,26 @@ import facebooklogo from '../../assets/logos_facebook.png';
 import googlelogo from '../../assets/google-original.png';
 import investorImg from '../../assets/investorImg.png'
 import ownerImg from '../../assets/ownerImg.png'
+import person from '../../assets/person.png'
+import correct from '../../assets/correct.png'
+
 
 const AuthRegister = () => {
+    const [firstName, setFirstName] = useState('');
+    const [lastName, setLastName] = useState('');
     const [category, setCategory] = useState('');
+    const [country, setCountry] = useState('');
     const [showForm, setShowForm] = useState(false);
     const [email, setEmail] = useState('');
     const [password, setPassword] = useState('');
     const [confirmPassword, setConfirmPassword] = useState('');
+    const [showConfirmPassword, setShowConfirmPassword] = useState(false);
     const [showPassword, setShowPassword] = useState(false);
+    const [gender, setGender] = useState('');
+    const [phone, setPhone] = useState('');
     const [error, setError] = useState('');
     const [isLoading, setIsLoading] = useState(false);
+    const [showModal, setShowModal] = useState(false);
     const navigate = useNavigate();
 
     const validateEmail = (email) => {
@@ -28,7 +38,9 @@ const AuthRegister = () => {
             setError('');
             setIsLoading(true);
 
-            if (!email || !password) {
+            if (!firstName || !lastName || !category || !country || !email || !password
+                || !confirmPassword
+            ) {
                 setError('Please fill in all fields');
                 setIsLoading(false);
                 return;
@@ -39,17 +51,27 @@ const AuthRegister = () => {
                 setIsLoading(false);
                 return;
             }
-
+            if (password !== confirmPassword) {
+                setError('Passwords do not match');
+                setIsLoading(false);
+                return;
+            }
             // Simulate API call
             setTimeout(() => {
+                console.log('Category:', category);
+                console.log('First Name:', firstName);
+                console.log('Last Name:', lastName);
                 console.log('Email:', email);
                 console.log('Password:', password);
+                console.log('Gender:', gender);
+                console.log('Phone:', phone);
+                console.log('Country:', country);
                 localStorage.setItem('isLoggedIn', 'true');
                 setIsLoading(false);
-                navigate('/');
+                setShowModal(true);
             }, 1000); // Replace with actual API call
         },
-        [email, password, navigate]
+        [email, password, navigate, firstName, lastName, category, country, confirmPassword]
     );
 
     const handleSocialLogin = (provider) => {
@@ -68,8 +90,8 @@ const AuthRegister = () => {
                         <h4 className={styles.loginText}>We help our users to make the right financial decisions</h4>
                     </div>
                     {/* left side */}
-                    <div className="col-8">
-                        <div className={styles.loginForm}>
+                    <div className="col-8 ms-auto">
+                        <div className={`${styles.loginForm} pb-4`}>
                             {!showForm ? (
                                 <div className={styles.categorySelection}>
                                     <h2>Select Your Role</h2>
@@ -110,11 +132,11 @@ const AuthRegister = () => {
                                         <div className="row mb-1">
                                             <div className="col-6">
                                                 <label htmlFor="firstName">First Name</label>
-                                                <input type="text" className="form-control" id="firstName" placeholder="First Name" />
+                                                <input type="text" value={firstName} onChange={(e) => setFirstName(e.target.value)} className="form-control" id="firstName" placeholder="First Name" />
                                             </div>
                                             <div className="col-6">
                                                 <label htmlFor="lastName">Last Name</label>
-                                                <input type="text" className="form-control" id="lastName" placeholder="Last Name" />
+                                                <input type="text" value={lastName} onChange={(e) => setLastName(e.target.value)} className="form-control" id="lastName" placeholder="Last Name" />
                                             </div>
                                         </div>
 
@@ -127,14 +149,14 @@ const AuthRegister = () => {
                                             {/* Country */}
                                             <div className="col-6">
                                                 <label htmlFor="country" className="form-label">Country</label>
-                                                <input type="text" className="form-control" id="country" placeholder="🇪🇬 EGYPT" />
+                                                <input type="text" value={country} onChange={(e) => setCountry(e.target.value)} className="form-control" id="country" placeholder="🇪🇬 EGYPT" />
                                             </div>
                                         </div>
                                         <div className="row mb-1">
                                             {/* Gender */}
                                             <div className='col-6'>
                                                 <label htmlFor="gender">Gender</label>
-                                                <select className="form-control" id="gender">
+                                                <select className="form-control" value={gender} onChange={(e) => setGender(e.target.value)} id="gender">
                                                     <option value="">Select Gender</option>
                                                     <option value="male">Male</option>
                                                     <option value="female">Female</option>
@@ -152,11 +174,10 @@ const AuthRegister = () => {
                                                         <option value="+20">EGY</option>
                                                         <option value="+91">IND</option>
                                                     </select>
-                                                    <input type="tel" className="form-control" id="phone" placeholder="+20100000" />
+                                                    <input type="tel" className="form-control" id="phone" value={phone} onChange={(e) => setPhone(e.target.value)} placeholder="+20100000" />
                                                 </div>
                                             </div>
                                         </div>
-
 
                                         {/* Password */}
                                         <div className="mb-1 position-relative">
@@ -187,7 +208,7 @@ const AuthRegister = () => {
                                             <label htmlFor="confirmPassword" className="form-label">Confirm Password</label>
                                             <div className="input-group">
                                                 <input
-                                                    type={showPassword ? "text" : "password"}
+                                                    type={showConfirmPassword ? "text" : "password"}
                                                     className="form-control rounded-end"
                                                     id="confirmPassword"
                                                     placeholder="Confirm Password"
@@ -197,12 +218,12 @@ const AuthRegister = () => {
                                                 />
                                                 <span
                                                     className={styles.inputGroupText}
-                                                    onClick={() => setShowPassword(!showPassword)}
+                                                    onClick={() => setShowConfirmPassword(!showConfirmPassword)}
                                                     role='button'
-                                                    aria-label={showPassword ? "Hide password" : "Show password"}
+                                                    aria-label={showConfirmPassword ? "Hide password" : "Show password"}
                                                     tabIndex={0}
-                                                    onKeyDown={(e) => e.key === 'Enter' && setShowPassword(!showPassword)}>
-                                                    {showPassword ? <FaEyeSlash /> : <FaEye />}
+                                                    onKeyDown={(e) => e.key === 'Enter' && setShowPassword(!showConfirmPassword)}>
+                                                    {showConfirmPassword ? <FaEyeSlash /> : <FaEye />}
                                                 </span>
                                             </div>
                                         </div>
@@ -225,6 +246,35 @@ const AuthRegister = () => {
                         </div>
                     </div>
                 </div>
+                {/* Custom modal */}
+                {showModal && (
+                    <div className={`${styles.overlay} ${showModal ? styles.open : ''}`}
+                        role="dialog"
+                        aria-labelledby="successModalLabel"
+                    // ref={modalRef}
+                    >
+                        <div className={styles.modalContent}>
+                            <div className={styles.modalBody}>
+                                <div className={styles.modalIcon}>
+                                    <img src={person} alt="person" className={styles.person} />
+                                    <img src={correct} alt="correct" className={styles.correct} />
+                                </div>
+                                <h3 id="successModalLabel">Your account has been created</h3>
+                                <p>Are you ready to manage your own FundX?</p>
+                                <button
+                                    type="button"
+                                    className={`${styles.loginBtn} ${styles.continueModalBtn}`}
+                                    onClick={() => {
+                                        setShowModal(false);
+                                        navigate('/')
+                                    }}
+                                >
+                                    Continue
+                                </button>
+                            </div>
+                        </div>
+                    </div>
+                )}
             </div>
         </div>
     )
